# openapi-spec.yaml
openapi: 3.0.3
info:
  title: Manus AI API
  description: |
    A comprehensive, versatile AI system built with NestJS that provides specialized AI agents capable of performing various tasks through real AI integration (OpenAI + LangChain), background processing, and system automation.

    ## Key Features
    - **9 Specialized AI Agents** with real AI integration (no mock data)
    - **Background Task Processing** with BullMQ and Redis queues
    - **Session Persistence** for offline capability and long-running tasks
    - **Real-time AI Interactions** with streaming responses via LangChain
    - **System Tool Integration** (file operations, browser automation, shell commands)

    ## Technologies Used
    - **OpenAI + LangChain**: Real AI integration with proper streaming and tool binding
    - **NestJS**: Progressive Node.js framework
    - **BullMQ**: Advanced Redis-based queue system
    - **MongoDB**: NoSQL database with Prisma ORM
    - **TypeScript**: Strongly typed development

  version: 1.0.0
  contact:
    name: Manus AI Team
    email: support@manus.im
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api/v1
    description: Development server
  - url: https://api.manus.im/api/v1
    description: Production server
paths:
...
  /agents/{agentName}:
    get:
      summary: Get specific agent details
      description: Returns detailed information about a specific AI agent
      tags:
        - Agents
      parameters:
        - name: agentName
          in: path
          required: true
          description: Name of the agent
          schema:
            type: string
            enum:
              - artistic_inspiration_agent
              - career_planning_agent
              - customer_service_agent
              - email_management_agent
              - learning_platform_agent
              - stress_management_agent
              - recommendation_agent
              - video_editing_agent
              - general_assistant_agent
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDetails'
        '404':
          description: Agent not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /agents/{agentName}/chat:
    post:
      summary: Chat with a specific AI agent
      description: Send a message to a specific agent and receive an AI-generated response using OpenAI + LangChain
      tags:
        - Agents
        - Chat
      parameters:
        - name: agentName
          in: path
          required: true
          description: Name of the agent to chat with
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatRequest'
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /agents/tools/execute:
    post:
      summary: Execute a specific AI tool
      description: Execute any specialized AI tool with parameters using LangChain integration
      tags:
        - Tools
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolExecuteRequest'
      responses:
        '200':
          description: Tool execution result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolExecuteResponse'
  /agents/tools/list:
    get:
      summary: Get list of all available AI tools
      description: Retrieve a comprehensive list of all specialized AI tools with their capabilities and requirements
      tags:
        - Tools
      parameters:
        - name: category
          in: query
          required: false
          description: Filter tools by category
          schema:
            type: string
            enum: ['web', 'content', 'analysis', 'automation', 'creative', 'productivity']
      responses:
        '200':
          description: List of available tools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ToolInfo'
  /agents/tools/{toolName}:
    get:
      summary: Get detailed information about a specific AI tool
      description: Get comprehensive details about a specific tool including usage examples and parameter specifications
      tags:
        - Tools
      parameters:
        - name: toolName
          in: path
          required: true
          description: Name of the tool to get information about
          schema:
            type: string
            example: 'search_web'
      responses:
        '200':
          description: Detailed tool information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolDetailInfo'
        '404':
          description: Tool not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Agent:
      type: object
      properties:
        name:
          type: string
          description: Agent identifier name
        description:
          type: string
          description: Agent description and capabilities
        tools:
          type: array
          items:
            type: string
          description: List of tools available to this agent
        category:
          type: string
          description: Agent category
    AgentDetails:
      allOf:
        - $ref: '#/components/schemas/Agent'
        - type: object
          properties:
            systemPrompt:
              type: string
              description: System prompt used by the agent
            settings:
              type: object
              properties:
                temperature:
                  type: number
                  format: float
                  description: AI model temperature setting
                maxTokens:
                  type: integer
                  description: Maximum tokens for AI responses
    ChatRequest:
      type: object
      required:
        - message
        - userId
      properties:
        message:
          type: string
          description: The message to send to the agent
        userId:
          type: string
          description: User identifier
        conversationHistory:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
          description: Previous conversation messages
        conversationId:
          type: string
          description: Conversation identifier
    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
    ChatResponse:
      type: object
      properties:
        agent:
          type: string
          description: Agent that processed the message
        response:
          type: string
          description: AI-generated response using OpenAI + LangChain
        toolsUsed:
          type: array
          items:
            type: string
          description: Tools used during response generation
        timestamp:
          type: string
          format: date-time
        conversationId:
          type: string
    ToolExecuteRequest:
      type: object
      required:
        - toolName
        - parameters
        - userId
      properties:
        toolName:
          type: string
          description: Name of the tool to execute
        parameters:
          type: object
          description: Parameters for tool execution
        userId:
          type: string
          description: User identifier
    ToolExecuteResponse:
      type: object
      properties:
        result:
          type: string
          description: Tool execution result
        toolName:
          type: string
        executionTime:
          type: number
          description: Execution time in milliseconds
        timestamp:
          type: string
          format: date-time
    ToolInfo:
      type: object
      properties:
        name:
          type: string
          description: Tool name
          example: 'search_web'
        description:
          type: string
          description: Tool description
          example: 'Search the web for information'
        category:
          type: string
          description: Tool category
          example: 'web'
        parameters:
          type: object
          description: Required parameters schema
          example:
            query:
              type: string
              required: true
              description: Search query
        agentRequired:
          type: string
          description: Required agent for this tool
          example: 'general_assistant_agent'
    ToolDetailInfo:
      allOf:
        - $ref: '#/components/schemas/ToolInfo'
        - type: object
          properties:
            examples:
              type: array
              items:
                type: object
                properties:
                  title:
                    type: string
                  parameters:
                    type: object
                  expectedOutput:
                    type: string
              description: Usage examples
            capabilities:
              type: array
              items:
                type: string
              description: Tool capabilities
              example: ['web_search', 'content_analysis', 'link_extraction']
            limitations:
              type: array
              items:
                type: string
              description: Tool limitations
            version:
              type: string
              description: Tool version
            lastUpdated:
              type: string
              format: date-time
              description: Last update timestamp
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Error type
        statusCode:
          type: integer
          description: HTTP status code
        timestamp:
          type: string
          format: date-time
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []
...